+++ module/Application/src/Application/Entity/Db/IndividuCompl.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace Application\Entity\Db;
+
+use UnicaenApp\Entity\HistoriqueAwareInterface;
+use UnicaenApp\Entity\HistoriqueAwareTrait;
+
+class IndividuCompl implements HistoriqueAwareInterface {
+    use HistoriqueAwareTrait;
+
+    /** @var int $id */
+    private $id;
+    /** @var Individu $individu */
+    private $individu;
+    /** @var string $email */
+    private $email;
+
+    /**
+     * @return int
+     */
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    /**
+     * @return Individu|null
+     */
+    public function getIndividu(): ?Individu
+    {
+        return $this->individu;
+    }
+
+    /**
+     * @param Individu $individu
+     * @return IndividuCompl
+     */
+    public function setIndividu(Individu $individu): IndividuCompl
+    {
+        $this->individu = $individu;
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getEmail(): ?string
+    {
+        return $this->email;
+    }
+
+    /**
+     * @param string $email
+     * @return IndividuCompl
+     */
+    public function setEmail(string $email): IndividuCompl
+    {
+        $this->email = $email;
+        return $this;
+    }
+
+}