+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Journal des modifications
 -----
 - Correction de typos dans mail de feu vert de la soutenance
 - Ajout de redirection de mail lorsque certains mails n'ont pas de destinataire "ATTENTION MAIL NON DÉLIVRÉ".
+- Compléments d'individu : mise en place des éléments de base 
 
 4.0.3
 -----